third_party/blink/web_tests/platform/generic/netinfo/estimate-basic-operation-expected.txt

Tests the basic operation of NetInfo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.connection is defined.
PASS navigator.connection.type is defined.
PASS navigator.connection.saveData is defined.
PASS navigator.connection.downlinkMax is defined.
PASS navigator.connection.effectiveType is defined.
PASS navigator.connection.rtt is defined.
PASS navigator.connection.downlink is defined.
PASS typeof connection.type is "string"
PASS connection.type is initialType
PASS connection.saveData is false
PASS connection.downlinkMax is initialDownlinkMax
PASS connection.effectiveType is newEffectiveType
PASS connection.rtt is newRtt
PASS connection.downlink is newDownlinkMax
PASS typeof connection.type is "string"
PASS connection.type is newConnectionType
PASS connection.saveData is false
PASS connection.downlinkMax is newDownlinkMax
PASS connection.effectiveType is newEffectiveType
PASS connection.rtt is newRtt
PASS connection.downlink is newDownlinkMax
PASS successfullyParsed is true

TEST COMPLETE
